% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{fileConversion}
\alias{fileConversion}
\title{Encode a file to PixPack PNG or decode a PixPack PNG back to original file.}
\usage{
fileConversion(path)
}
\arguments{
\item{path}{Input file path (character scalar).}
}
\value{
Output artifact path (character scalar).
}
\description{
Encode a file to PixPack PNG or decode a PixPack PNG back to original file.
}
